:bug: fix: 에상치 못한 오류로 인해 수정 사항들 다시 돌려놓기
