+++ src/main/java/lotto/model/LottoFactory.java
@@ -4,7 +4,6 @@ import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
-import lotto.Lotto;
 import lotto.utilities.Random;
 
 public class LottoFactory {
