+++ src/main/java/lotto/model/Lottos.java
@@ -2,8 +2,6 @@ package lotto.model;
 
 import java.util.List;
 
-import lotto.Lotto;
-
 public class Lottos {
     private final List<Lotto> lottos;
